# apps/architectural-service/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Architectural Service API",
...
    "contact": {
      "name": "DesignSynapse Support",
      "url": "https://designsynapse.com/support",
      "email": "support@designsynapse.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://designsynapse.com/license"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/designs/{design_id}/drawings": {
      "post": {
        "tags": [
          "drawings"
        ],
        "summary": "Upload a drawing",
        "description": "Upload an architectural drawing file to a design document",
        "operationId": "upload_drawing_api_v1_designs__design_id__drawings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "design_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Design Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_drawing_api_v1_designs__design_id__drawings_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrawingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "drawings"
        ],
        "summary": "List drawings for a design",
        "description": "List all drawings associated with a design document",
        "operationId": "list_drawings_api_v1_designs__design_id__drawings_get",
        "parameters": [
          {
            "name": "design_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Design Id"
            }
          },
          {
            "name": "drawing_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DrawingType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by drawing type",
              "title": "Drawing Type"
            },
            "description": "Filter by drawing type"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrawingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drawings/{drawing_id}": {
      "get": {
        "tags": [
          "drawings"
        ],
        "summary": "Get a drawing",
        "description": "Retrieve a specific drawing with detailed information",
        "operationId": "get_drawing_api_v1_drawings__drawing_id__get",
        "parameters": [
          {
            "name": "drawing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Drawing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrawingDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "drawings"
        ],
        "summary": "Delete a drawing",
        "description": "Delete a drawing and its associated file",
        "operationId": "delete_drawing_api_v1_drawings__drawing_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drawing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Drawing Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_drawing_api_v1_designs__design_id__drawings_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Drawing file to upload"
          },
          "drawing_type": {
            "$ref": "#/components/schemas/DrawingType",
            "description": "Type of architectural drawing"
          },
          "scale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale",
            "description": "Drawing scale"
          },
          "sheet_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Number",
            "description": "Sheet number"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Additional metadata as JSON string"
          }
        },
        "type": "object",
        "required": [
          "file",
          "drawing_type"
        ],
        "title": "Body_upload_drawing_api_v1_designs__design_id__drawings_post"
      },
...
      "DrawingDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Drawing ID"
          },
          "design_id": {
            "type": "string",
            "format": "uuid",
            "title": "Design Id",
            "description": "Associated design ID"
          },
          "design_version": {
            "type": "string",
            "title": "Design Version",
            "description": "Design version when drawing was added"
          },
          "drawing_type": {
            "$ref": "#/components/schemas/DrawingType",
            "description": "Type of architectural drawing"
          },
          "file_url": {
            "type": "string",
            "title": "File Url",
            "description": "URL to access the drawing file"
          },
          "file_size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "File Size",
            "description": "File size in bytes"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type",
            "description": "MIME type of the file"
          },
          "scale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale",
            "description": "Drawing scale"
          },
          "sheet_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Number",
            "description": "Sheet number in drawing set"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional drawing metadata"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By",
            "description": "User ID who uploaded the drawing"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Upload timestamp"
          },
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Name",
            "description": "Original file name"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url",
            "description": "URL to thumbnail preview"
          }
        },
        "type": "object",
        "required": [
          "id",
          "design_id",
          "design_version",
          "drawing_type",
          "file_url",
          "file_size",
          "mime_type",
          "created_by",
          "created_at"
        ],
        "title": "DrawingDetailResponse",
        "description": "Detailed response schema for drawing with additional information.",
        "example": {
          "created_at": "2024-01-15T10:30:00Z",
          "created_by": "770e8400-e29b-41d4-a716-446655440000",
          "design_id": "550e8400-e29b-41d4-a716-446655440000",
          "design_version": "1.0",
          "drawing_type": "floor_plan",
          "file_size": 2048576,
          "file_url": "https://storage.example.com/drawings/990e8400.pdf",
          "id": "990e8400-e29b-41d4-a716-446655440000",
          "metadata": {
            "level": "Ground Floor",
            "revision": "A"
          },
          "mime_type": "application/pdf",
          "scale": "1/4\"=1'",
          "sheet_number": "A-101"
        }
      },
      "DrawingListResponse": {
        "properties": {
          "drawings": {
            "items": {
              "$ref": "#/components/schemas/DrawingResponse"
            },
            "type": "array",
            "title": "Drawings",
            "description": "List of drawings"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of drawings"
          }
        },
        "type": "object",
        "required": [
          "drawings",
          "total"
        ],
        "title": "DrawingListResponse",
        "description": "Response schema for list of drawings."
      },
      "DrawingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Drawing ID"
          },
          "design_id": {
            "type": "string",
            "format": "uuid",
            "title": "Design Id",
            "description": "Associated design ID"
          },
          "design_version": {
            "type": "string",
            "title": "Design Version",
            "description": "Design version when drawing was added"
          },
          "drawing_type": {
            "$ref": "#/components/schemas/DrawingType",
            "description": "Type of architectural drawing"
          },
          "file_url": {
            "type": "string",
            "title": "File Url",
            "description": "URL to access the drawing file"
          },
          "file_size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "File Size",
            "description": "File size in bytes"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type",
            "description": "MIME type of the file"
          },
          "scale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale",
            "description": "Drawing scale"
          },
          "sheet_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Number",
            "description": "Sheet number in drawing set"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional drawing metadata"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By",
            "description": "User ID who uploaded the drawing"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Upload timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "design_id",
          "design_version",
          "drawing_type",
          "file_url",
          "file_size",
          "mime_type",
          "created_by",
          "created_at"
        ],
        "title": "DrawingResponse",
        "description": "Response schema for drawing.",
        "example": {
          "created_at": "2024-01-15T10:30:00Z",
          "created_by": "770e8400-e29b-41d4-a716-446655440000",
          "design_id": "550e8400-e29b-41d4-a716-446655440000",
          "design_version": "1.0",
          "drawing_type": "floor_plan",
          "file_size": 2048576,
          "file_url": "https://storage.example.com/drawings/990e8400.pdf",
          "id": "990e8400-e29b-41d4-a716-446655440000",
          "metadata": {
            "level": "Ground Floor",
            "revision": "A"
          },
          "mime_type": "application/pdf",
          "scale": "1/4\"=1'",
          "sheet_number": "A-101"
        }
      },
      "DrawingType": {
        "type": "string",
        "enum": [
          "floor_plan",
          "elevation",
          "section",
          "site_plan",
          "detail",
          "roof_plan",
          "reflected_ceiling_plan",
          "3d_view"
        ],
        "title": "DrawingType",
        "description": "Architectural drawing types."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}